refactor: remove colorOverlap
