+++ __tests__/unit/utils/colorBlend.test.ts
@@ -10,6 +10,115 @@ describe('Color blend', () => {
     expect(colorToHex(colorBlend(colorTop, colorBottom)).toUpperCase()).toBe('#5B8FF9');
   });
 
+  test('transparent color overlap on white', () => {
+    const color1: Color = {
+      model: 'rgba',
+      value: {
+        r: 0,
+        g: 0,
+        b: 0,
+        a: 0,
+      },
+    };
+    const overlappedColor = colorBlend(color1, WHITE);
+    expect(overlappedColor).toStrictEqual(WHITE);
+  });
+
+  test('white overlap on transparent color', () => {
+    const color1: Color = {
+      model: 'rgba',
+      value: {
+        r: 0,
+        g: 0,
+        b: 0,
+        a: 0,
+      },
+    };
+    const overlappedColor = colorBlend(WHITE, color1);
+    expect(overlappedColor).toStrictEqual(WHITE);
+  });
+
+  test('color overlap on white', () => {
+    const color1: Color = {
+      model: 'rgba',
+      value: {
+        r: 0,
+        g: 0,
+        b: 0,
+        a: 0.8,
+      },
+    };
+    const overlappedColor = colorBlend(color1, WHITE);
+    expect(overlappedColor).toStrictEqual({
+      model: 'rgb',
+      value: {
+        r: 51,
+        g: 51,
+        b: 51,
+      },
+    });
+  });
+
+  test('color overlap on black', () => {
+    const color1: Color = {
+      model: 'rgba',
+      value: {
+        r: 123,
+        g: 224,
+        b: 156,
+        a: 0.6,
+      },
+    };
+    const black: Color = {
+      model: 'rgb',
+      value: {
+        r: 0,
+        g: 0,
+        b: 0,
+      },
+    };
+    const overlappedColor = colorBlend(color1, black);
+    expect(overlappedColor).toStrictEqual({
+      model: 'rgb',
+      value: {
+        r: 74,
+        g: 134,
+        b: 94,
+      },
+    });
+  });
+
+  test('color overlap on semi-transparent', () => {
+    const color1: Color = {
+      model: 'rgba',
+      value: {
+        r: 123,
+        g: 224,
+        b: 156,
+        a: 0.6,
+      },
+    };
+    const color2: Color = {
+      model: 'rgba',
+      value: {
+        r: 66,
+        g: 77,
+        b: 88,
+        a: 0.3,
+      },
+    };
+    const overlappedColor = colorBlend(color1, color2);
+    expect(overlappedColor).toStrictEqual({
+      model: 'rgba',
+      value: {
+        r: 114,
+        g: 200,
+        b: 145,
+        a: 0.72,
+      },
+    });
+  });
+
   test('darken', () => {
     expect(colorToHex(colorBlend(colorTop, colorBottom, 'darken')).toUpperCase()).toBe('#5B8FAA');
   });
