+++ demo/src/components/ColorComputation/index.tsx
@@ -1,13 +1,11 @@
 import React from 'react';
 import ColorBrighten from './ColorBrighten';
 import ColorDarken from './ColorDarken';
-import ColorOverlap from './ColorOverlap';
 import ColorBlend from './ColorBlend';
 
 const ColorComputation = () => {
   return (
     <>
-      <ColorOverlap />
       <ColorBlend />
       <ColorBrighten />
       <ColorDarken />
