+++ src/evaluators/colorDifference/index.ts
@@ -1,6 +1,6 @@
 import { ColorDifference } from '@src/types';
 import { WHITE } from '../../constant';
-import { colorOverlap } from '../../utils';
+import { colorBlend } from '../../utils';
 import { euclideanDistance } from './euclideanDistance';
 import { CIEDE2000 } from './CIEDE2000';
 import { contrastRatio } from './contrastRatio';
@@ -8,8 +8,8 @@ import { contrastRatio } from './contrastRatio';
 export const colorDifference: ColorDifference = (color1, color2, configuration = { measure: 'euclidean' }) => {
   const { measure = 'euclidean', backgroundColor = WHITE } = configuration;
   // If the color is semi-transparent, the color will be overlapped on a white background by default for comparison
-  const overlappedColor1 = colorOverlap(color1, backgroundColor);
-  const overlappedColor2 = colorOverlap(color2, backgroundColor);
+  const overlappedColor1 = colorBlend(color1, backgroundColor);
+  const overlappedColor2 = colorBlend(color2, backgroundColor);
   switch (measure) {
     case 'CIEDE2000':
       return CIEDE2000(overlappedColor1, overlappedColor2);
